Refactored HolidayService to be based on Organization instead.
